Update documentation and enable code metadata generation
