+++ docs/api/toc.yml
@@ -1,0 +1,4 @@
+﻿- name: REST API
+  href: rest.md
+- name: .NET
+  href: ../obj/api/dotnet/toc.yml
