+++ docs/filterConfig.yml
@@ -1,0 +1,12 @@
+apiRules:
+- exclude:
+    hasAttribute:
+      uid: Newtonsoft.Json.JsonIgnoreAttribute
+- exclude:
+    # inherited Object methods
+    uidRegex: ^System\.Object\..*$
+    type: Method
+- exclude:
+    # mentioning types from System.* namespace
+    uidRegex: ^System\..*$
+    type: Type
