+++ docs/templates/material/styles/main.css
@@ -11,6 +11,7 @@
     --card-box-shadow: 0 1px 2px 0 rgba(61, 65, 68, 0.06), 0 1px 3px 1px rgba(61, 65, 68, 0.16);
     --search-box-shadow: 0 1px 2px 0 rgba(41, 45, 48, 0.36), 0 1px 3px 1px rgba(41, 45, 48, 0.46);
     --transition: 350ms;
+    --hover: #eff7e0;
 }
 
 body {
@@ -322,4 +323,14 @@ pre {
     margin-top: 15px;
     box-shadow: var(--card-box-shadow);
     max-width: 700px;
+}
+
+#fields,
+#properties,
+#methods,
+#events {
+    font-weight: bold;
+    margin-top: 2em;
+    border-bottom: 8px solid var(--highlight-dark);
+    padding: 0 10px;
 }
