+++ src/API/Rest/InferenceRequestMetadata.cs
@@ -1,0 +1,58 @@
+/*
+ * Apache License, Version 2.0
+ * Copyright 2019-2020 NVIDIA Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Nvidia.Clara.Dicom.Common;
+using Nvidia.Clara.DicomAdapter.Common;
+using Nvidia.Clara.Platform;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Runtime.Serialization;
+using System.Text;
+
+namespace Nvidia.Clara.DicomAdapter.API.Rest
+{
+    /// <summary>
+    /// Represents the metadata associated with an inference request.
+    /// </summary>
+    /// <example>
+    /// <code>
+    /// {
+    ///     ...
+    ///     "inputMetadata" : {
+    ///         "details" : { ... }
+    ///     }
+    ///     ...
+    /// }
+    /// </code>
+    /// </example>
+    /// <remarks>
+    /// <para><c>details></c> is required.</para>
+    /// </remarks>
+    public class InferenceRequestMetadata
+    {
+        /// <summary>
+        /// Gets or sets the details of an inference request.
+        /// </summary>
+        [Required]
+        [JsonProperty(PropertyName = "details")]
+        public InferenceRequestDetails Details { get; set; }
+    }
+}
